Refactor the code for a simpler structure.
Overall the different clients were embedded inside the pmk pkg.
The clients created for qbert, resmgr and keystone are all very
interesting and other projects like airctl would benefit from importing
them. This refactor moves things around and promotes bunch of embedded
projects into first class pkgs for simpler organization of the code.

This is part 1 of the two part refactor I would like to do. The second
phase would deal with the logs and the constant but that is better
done as a seperate commit. Remember this doesn't change any
functionality, it helps with external and internal usage of the code.