More refactoring (#23)
* Refactor the code for a simpler structure.

Overall the different clients were embedded inside the pmk pkg.
The clients created for qbert, resmgr and keystone are all very
interesting and other projects like airctl would benefit from importing
them. This refactor moves things around and promotes bunch of embedded
projects into first class pkgs for simpler organization of the code.

This is part 1 of the two part refactor I would like to do. The second
phase would deal with the logs and the constant but that is better
done as a seperate commit. Remember this doesn't change any
functionality, it helps with external and internal usage of the code.

* Remove unnecessary arguments to PrepNode

A simple refactoring to remove extra parameter from the PrepNode call.
These parameters were not used and didn't belong there in the first
place.

* Fixing yum issue

* Adding github actions instead of make

* Create README.md

* Removing the log wrapper.

Removing the log wrapper, there is no need to create another wrapper.
Uber zap provides a way to Replace global logger. This makes it easier
for the caller (specifically external importers) to control how and
where the logs must be placed.

* Moving context form the pkg to cmd.

This move is to make sure:
* Pkg code is independent of the context (which is very pf9ctl
specific).
* Hide encoding decoding of password to the context code only
* Fix other bugs