Add --kube-version hidden flag to prep node
This commit adds a --kube-version arg to prep node command. This will be used in CAPI boot scripts to simplifying and improve the onboarding process.
Currently marking this as a hidden flag but if this turns out to be useful it can be made visible to the users as well.